# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "A RESTful API service for SPAC",
        "title": "SPAC REST API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
...
        "/map": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a map with specified properties",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "maps"
                ],
                "summary": "Create a new map",
                "parameters": [
                    {
                        "description": "Map creation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.CreateMapRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.CreateMapResponse"
                        }
                    }
                }
            }
        },
...
        "/map/{mapId}/elements": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Add multiple elements to an existing map",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "maps"
                ],
                "summary": "Add elements to a map",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Map ID",
                        "name": "mapId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Map elements to add",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.AddMapElementRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.MapElement": {
            "type": "object",
            "properties": {
                "X": {
                    "type": "integer"
                },
                "Y": {
                    "type": "integer"
                },
                "elementID": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "mapID": {
                    "type": "string"
                }
            }
        },
        "schemas.AddMapElementRequest": {
            "type": "object",
            "required": [
                "elements"
            ],
            "properties": {
                "elements": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.MapElement"
                    }
                }
            }
        },
...
        "schemas.CreateMapRequest": {
            "type": "object",
            "required": [
                "dimensions",
                "name",
                "thumbnail"
            ],
            "properties": {
                "dimensions": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "thumbnail": {
                    "type": "string"
                }
            }
        },
        "schemas.CreateMapResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}